Fixed bug of double char
